refactor: extract OrderDetail sections into components (DEBT-1 D1-C)
Mechanical extraction from pages/admin/OrderDetail.jsx (2,021 lines) into
components/orders/ per docs/plans/2026-06-12-god-file-decomposition.md:

- OrderWorkflowPanel.jsx (408): 4-step workflow panel + secondary order
  actions row; getOrderWorkflowSteps, step-state styling helpers, and the
  panel-only gating predicates (canCancelOrder/canCloseShort/canDeleteOrder/
  canGenerateInvoice/getShipBlockReason/canShipOrder) moved verbatim.
- LegacyFulfillmentBanner.jsx (137): LEGACY-1 amber banner + its confirm
  dialog + handleResolveLegacyFulfillment; resolution state is local now.
- OrderLineItemsTable.jsx (214): line-items table incl. inline edit state,
  handleSaveLineEdit/handleRemoveLine/canEditLines.
- OrderHeaderActions.jsx (33): header Refresh + Print Packing Slip buttons.
- orderWorkflowUtils.js (11): formatMoney + SHIPPED/UNCONFIRMED status sets
  shared by the page and the panel (own module so react-refresh stays clean).

Page is 1,369 lines and keeps data fetching, shared predicates, top-level
state, and modal orchestration (incl. ReleaseScheduleWizard wiring); state
flows down via explicit props. Zero behavior change; JSX/handlers moved
verbatim apart from handler->prop renames.

Verified: npm run test:unit 444/444 green unmodified, npm run build green,
eslint on touched files at pre-existing baseline (2 errors, issue #734).

Agent-Session: d1c-orderdetail
